Fix bug where you could remove points from early tiers
If you had enough points in a tree, you could remove points from earlier tiers
and create invalid trees. This was because only the total number of points in a
tree was stored, so to solve this we simply calculate the points in a tree every
time we need to know it.

Realistically, this whole thing could be re-written to be a bit cleaner.